chore(packages/cli): abstract the integration linter from a base linter
